+++ src/lib/tileinfo-importer/index.ts
@@ -1,0 +1,2 @@
+export { MetadataJSONImporter } from './metadata-importer'
+export { TileJSONImporter } from './tilejson-importer'
